more integration tests
