+++ lib/preflight.js
@@ -24,7 +24,7 @@ const { PreflightTest } = require('./preflighttest');
  *   console.log("Test completed in " + report.testTiming.duration + " milliseconds.");
  *   console.log(" It took " + report.networkTiming.connect.duration + " milliseconds to connect");
  *   console.log(" It took " + report.networkTiming.media.duration + " milliseconds to receive media");
- *   console.log(" Your network score was: " report.stats.networkQuality); // only available for group rooms.
+ *   console.log(" Your network score was: " + report.stats.networkQuality); // only available for group rooms.
  * });
  *
  * preflightTest.on('failed', function(error) {
