+++ lib/preflighttest.js
@@ -139,12 +139,6 @@ function runPreflightTest(publisherToken, subscriberToken, options, preflightTes
   let connectTiming = null;
   let mediaTiming = null;
 
-  const checkIfStopped = () => {
-    if (preflightTest._stopped) {
-      throw new Error('stopped');
-    }
-  };
-
   // collects stats at regular interval
   function collectStats(collectedStats) {
     return subscriberRoom._signaling.getStats().then(responses => {
@@ -249,62 +243,114 @@ function runPreflightTest(publisherToken, subscriberToken, options, preflightTes
     });
   }
 
-  createLocalTracks()
-    .then(tracks => {
-      localTracks = tracks;
-      checkIfStopped();
-      preflightTest.emit('progress', PreflightProgress.mediaAcquired);
-      return connect(publisherToken, options);
+  /**
+   * returns a promise to executes given step
+   * rejects the return promise if
+   * a) preflight is stopped.
+   * b) subscriber or publisher disconnects
+   * c) step does not complete in reasonable time.
+   * @param {function} step - function to execute
+   * @param {string} stepName - name for the step
+   */
+  function executePreflightStep(stepName, step) {
+    const MAX_STEP_DURATION = testDuration + 10 * SECOND;
+    if (preflightTest._stopped) {
+      throw new Error('stopped');
+    }
+
+    if (subscriberRoom && subscriberRoom.state === 'disconnected') {
+      throw new Error('subscriber disconnected unexpectedly');
+    }
+
+    if (publisherRoom && publisherRoom.state === 'disconnected') {
+      throw new Error('publisher disconnected unexpectedly');
+    }
+
+    const stepPromise = Promise.resolve().then(step);
+    let timer = null;
+    const timeoutPromise = new Promise((_resolve, reject) => {
+      timer = setTimeout(() => {
+        reject(new Error(`Timed out waiting for : ${stepName}`));
+      }, MAX_STEP_DURATION);
+    });
+    return Promise.race([timeoutPromise, stepPromise]).finally(() => {
+      clearTimeout(timer);
+    });
+  }
+
+  return Promise.resolve()
+    .then(() => {
+      return executePreflightStep('acquire media', () => {
+        return createLocalTracks();
+      });
+    }).then(tracks => {
+      return executePreflightStep('connect publisher', () => {
+        localTracks = tracks;
+        preflightTest.emit('progress', PreflightProgress.mediaAcquired);
+        return connect(publisherToken, options);
+      });
     }).then(room => {
-      publisherRoom = room;
-      checkIfStopped();
-      connectTiming = new TimeMeasurement();
-      mediaTiming = new TimeMeasurement();
-      return connect(subscriberToken, Object.assign(options, { name: room.sid }));
+      return executePreflightStep('connect subscriber', () => {
+        publisherRoom = room;
+        connectTiming = new TimeMeasurement();
+        mediaTiming = new TimeMeasurement();
+        return connect(subscriberToken, Object.assign(options, { name: room.sid }));
+      });
     }).then(room => {
-      subscriberRoom = room;
-      connectTiming.stop();
-      checkIfStopped();
-      preflightTest.emit('progress', PreflightProgress.connected);
-      // hook up for track started.
-      trackStartListener = listenForTrackStarted(subscriberRoom, 2);
-      return participantsConnected(publisherRoom);
+      return executePreflightStep('publisher sees subscriber', () => {
+        subscriberRoom = room;
+        connectTiming.stop();
+        preflightTest.emit('progress', PreflightProgress.connected);
+        trackStartListener = listenForTrackStarted(subscriberRoom, 2);
+        return participantsConnected(publisherRoom);
+      });
     }).then(() => {
-      checkIfStopped();
-      return participantsConnected(subscriberRoom);
+      return executePreflightStep('subscriber sees publisher', () => {
+        return participantsConnected(publisherRoom);
+      });
     }).then(() => {
-      checkIfStopped();
-      if (subscriberRoom.sid !== publisherRoom.sid) {
-        throw new Error('wrong room tokens. Ensure that both token join to the same room');
-      }
-      if (subscriberRoom.participants.size !== 1) {
-        throw new Error('unexpected participant found in the room');
-      }
-      preflightTest.emit('progress', PreflightProgress.remoteConnected);
+      return executePreflightStep('validate rooms', () => {
+        if (subscriberRoom.sid !== publisherRoom.sid) {
+          throw new Error('wrong room tokens. Ensure that both token join to the same room');
+        }
+        if (subscriberRoom.participants.size !== 1) {
+          throw new Error('unexpected participant found in the room');
+        }
+        preflightTest.emit('progress', PreflightProgress.remoteConnected);
+      });
     }).then(() => {
-      checkIfStopped();
-      return publisherRoom.localParticipant.publishTracks(localTracks);
+      return executePreflightStep('publish tracks', () => {
+        return publisherRoom.localParticipant.publishTracks(localTracks);
+      });
     }).then(() => {
-      checkIfStopped();
-      preflightTest.emit('progress', PreflightProgress.mediaPublished);
+      return executePreflightStep('emit mediaPublished', () => {
+        preflightTest.emit('progress', PreflightProgress.mediaPublished);
+      });
     }).then(() => {
-      checkIfStopped();
-      return tracksSubscribed(subscriberRoom, 2);
+      return executePreflightStep('wait for tracks to be subscribed', () => {
+        return tracksSubscribed(subscriberRoom, 2);
+      });
     }).then(() => {
-      checkIfStopped();
-      preflightTest.emit('progress', PreflightProgress.mediaSubscribed);
+      return executePreflightStep('emit mediaSubscribed', () => {
+        preflightTest.emit('progress', PreflightProgress.mediaSubscribed);
+      });
     }).then(() => {
-      checkIfStopped();
-      return trackStartListener.trackStarted();
+      return executePreflightStep('wait for tracks to start', () => {
+        return trackStartListener.trackStarted();
+      });
     }).then(() => {
-      checkIfStopped();
-      mediaTiming.stop();
-      preflightTest.emit('progress', PreflightProgress.mediaStarted);
+      return executePreflightStep('emit mediaStarted', () => {
+        mediaTiming.stop();
+        preflightTest.emit('progress', PreflightProgress.mediaStarted);
+      });
     }).then(() => {
-      return collectStatsPeriodically(testDuration);
+      return executePreflightStep('collect stats', () => {
+        return collectStatsPeriodically(testDuration);
+      });
     }).then(collectedStats => {
-      checkIfStopped();
-      return generatePreflightReport(collectedStats);
+      return executePreflightStep('generate report', () => {
+        return generatePreflightReport(collectedStats);
+      });
     }).then(report => {
       preflightTest.emit('completed', report);
     }).catch(error => {
